requirements.txt
altgraph==0.17.4
click==8.1.7
macholib==1.16.3
mutagen==1.47.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
six==1.16.0
text-unidecode==1.3
tzdata==2024.1
